op type fix && input control added
this should compile now finally, am not sure yet thoug.
the test produces this error " Op type not registered 'LinearEq' in
binary running on aca-project. Make sure the Op and Kernel are
registered in the binary running in this process.
" when registering XLA operation.

best guess it's a mapping problem